requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
dlib==19.24.2
idna==3.6
imageio==2.33.1
IsobarImg==0.4.2
lazy_loader==0.3
networkx==3.2.1
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
tifffile==2023.12.9
urllib3==2.2.0
